requirements.txt
fastapi[standart]
uvicorn[standart]
pydantic[email]
pymysql
sqlalchemy
python-dotenv
httpx
pytest
python-jose[cryptography]
passlib[bcrypt]